requirements.txt
fastapi
uvicorn
python-dotenv
motor
pymongo
python-multipart